Fix engagement_rate column type: Float -> Numeric(5,2) to match ORM model

+++ backend/alembic/versions/add_social_columns_to_prospects.py
@@ -140,7 +140,7 @@ def upgrade():
     # Add engagement_rate column
     if 'engagement_rate' not in existing_columns:
         op.add_column('prospects',
-            sa.Column('engagement_rate', sa.Float(), nullable=True)
+            sa.Column('engagement_rate', sa.Numeric(5, 2), nullable=True)
         )
         print("✅ Added engagement_rate column")
     else:
